Move rediscover suggester to git machete panel level
